event_hub/requirements.txt
fastapi
uvicorn[standard]
motor
strawberry-graphql[fastapi]
redis
pydantic